backend/requirements.txt
fastapi==0.111.1
uvicorn[standard]==0.23.2
sqlalchemy==2.0.23
pdfplumber==0.10.1
python-dotenv==1.0.0
langchain==0.1.0
openai==1.33.1
pydantic==2.2.0
